refactor: BaseTimeEntity 위치 common dir으로 변경
